Calendar: Add snooze and dismiss button for alert event
In Android Sources, there is no snooze button on the alert
event pop-up box and no dismiss button if the event only
alert in status bar. It's a new requirement.

Add snooze button on the alert event pop-up box and dismiss
button on the alert event in status bar.

CRs-Fixed: 395367

Change-Id: Ie751cace1666666ae5b719416a07df5ef8355ebd