Calendar: allow for custom event info text size
The new standard for the devices we maintain is to show events while in month
mode in Calendar (although it should be noted that this is completely optional
and done per-device via overlay). While we can adjust the event title text
size per-device (to make up for various screen sizes), the event info font
remains unchanged and causes sizes to look disproportional. This commit
addresses that.

Change-Id: I28ee32a540b1237ac832d42842470bf7ee44874b
Signed-off-by: Luis Cruz <ljc2491@gmail.com>